refact: move some code into authz.discovery
